extract ReadReceiverInfo
